completed till file upload

+++ controller/posts_controller.js
@@ -7,6 +7,16 @@ module.exports.create= async function(req, res){
                 content: req.body.content,
                 user: req.user._id
             });
+
+            if(req.xhr){
+                return res.status(200).json({
+                    data: {
+                        post: post
+                    },
+                    message: "Post created!"
+                })
+            }
+            
             req.flash('success', 'post created successfully!!');
             return res.redirect('back');
         }catch(err){
@@ -38,6 +48,15 @@ module.exports.destroy = async function(req, res){
 
     if(post.user == req.user.id){
         post.remove();
+
+        if(req.xhr){
+            return res.status(200).json({
+                data: {
+                    post_id: req.params.id
+                },
+                message: "Post deleted"
+            });
+        }
 
         await Comment.deleteMany({post: req.params.id});
         req.flash('success', 'post deleted successfully!!')
